Pictures and window icon
I added some folders for pictures, along with a couple pictures. Still
need to make the internal views for the ships. Also need to find more
ships in general.

I also added in an icon, and the code to display the icon was added to
main.py

The icon still has watermarks... :/
I'll fix it later.